requirements.txt
Pillow==9.2.0
Pyrogram==1.2.20
PyYAML==6.0
